Run CI tests with consistent node version
The version of node that is installed to the GitHub Action runner hosts
is a moving target which means that we'll have non-deterministic
artifact builds.

This PR corrects that by defining an [asdf-vm](https://asdf-vm.com/#/)
`.tool-versions` file with which version of node to use. In the CI
workflow where we run the tests and try to build the `ncc` artifact, we
reference the `.tool-versions` so that we can always have a consistent
and predictable version of node to build the artifact.

If there's a discrepancy in what gets built versus what is checked in,
we'll automatically push a new commit to make sure we always have a
deterministic build.

Lastly, everything was formatted with the yaml language server
formatter.